Simplifies Initialization and adds customization
This should simplify the initialization process for ocm-container as
well as add the ability for a user to further customize their
environment through the config file.

The main benefit of moving the config file to a neutral location is that
the config file can now be stored in a users' dotfiles and however they
have those set to be backed up, instead of being stored within the main
repo.