Adds `create_cluster` script to generate OCM clusters
This commit adds a Python script to create clusters easily, generating
cluster names and using the latest version available in OCM, or allowing
the user to specify a name and cluster version.

This allows for creating custom or generic clusters with random names
containing the user's OCM id for easy identification.  Cluster can be
created easily with 0 arguments if desired.

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>