Adjusts sre-login script to accept multiple arguments
Adjusts the sre-login script to accept multiple arguments:

    -h prints help message
    -d prints debug information (bash xtrace)
    -c cluster name
    -i OCM identity

The added help message and debug will assist new users or folks having
issues tracking down where their login problems are occuring, and the
OCM identiy argument will allow setting of any ID needed to login to the
cluster.

This comes at the expense of needed to pass `-c cluster` rather than
just the bare cluster name, but I think this is more clear usage, and
less prone to issues.

I'm happy to adjust to accept the old usage as well, or to add a
deprecation message, if that's desired.

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>