pause hive sync (#37)
* Adds utility to list available scripts in ocm-container

Adds a script utility `list-utils` that will list extra utils installed
in ocm-contianer, and print a description of what they do.

The description is pulled from a comment in the individual script
formatted as: `# OCM_CONTAINER_DOC: <description>`

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>

* Adds script to pause hive sync

Adds a script to pause hive sync for a cluster. Must be logged into the
hive that controls the cluster you wish to pause syncing on.

Usage: `pause-hive-sync -c CLUSTER_NAME`

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>

* Update container-setup/utils/bin/pause-hive-sync

+1, thanks!

Co-authored-by: Ron Green <11993626+georgettica@users.noreply.github.com>

* Update container-setup/utils/bin/pause-hive-sync

This was probably a copy/paste issue when moved from another part of the script.  Yours is 100% better, thanks!

Co-authored-by: Ron Green <11993626+georgettica@users.noreply.github.com>

* remove unecessary echo -e flag

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>

* Adds toggle; handles errors

* Adds a toggle to pause or unpause syncing from hive
* Exits more correctly for errors, writes output to stderr

CC @mrbarge

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>

Co-authored-by: Ron Green <11993626+georgettica@users.noreply.github.com>