Adds auto-mapping for cluster console port and a wrapper for use
Adds an auto mapping option for the cluster console port and a script to launch it.  Also includes a way to disable the automatic port mapping if desired.

Modifes the ocm-contianer.sh script to extract the randomized port mapping and inject it into the container via file.

Adds a wrapper to `ocm backplane console` via bash function, to make the output use the mapped port and allow for a copy/paste or clickable link to open in the browser outside the container.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>