Moves chmod of binaries to before final image
This moves the final chmod of all the binaries from one spot in the
final image build to the upstream build images.

Running chmod within an image creates a new copy of the binary as a
layer over the other, so running it in the final image doubles the size
used by each binary, whereas running it in the builder images prevents
that.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>