add a minimal target for ocm-container-minimal
This PR reorders the Container file to add an ocm-contianer-minimal
build target consisting of the base backplane-tools install.

This includes a couple of other small reorder changes that should allow
for more caching during local builds for quicker and easier testing,
particularly of the backplane-tools install steps and the image layers
built before it.

The Makefile has been updated to add a `build-minimal` target.

The main ocm-container image, build pipeline, and the rest of the
Makefile will be unchanged by this PR.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>