Replace 0.0.0.0 with 127.0.0.1 in blackbox-exporter output
Replaces the `0.0.0.0` with `127.0.0.1` in the blackbox-exporter output.
This does not functionally change anything, but matches the external
expectations that 127.0.0.1 is the port in use, to avoid confusion. The
internal address for the port forward has to be 0.0.0.0 or it will be
unavailable outside of the container.  However, the external port map is
still 127.0.0.1, and only available on the local machine, whether
accessed via 127.0.0.1 or 0.0.0.0 in the browser.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>