Reorder bashrc.d by type; skip automation if launched without bin
This PR reorders and reorganizes the bashrc.d files, placing pure
aliases and functions low (0-9) and labeling them with "libs", exports
and sourcing followign that (10-19) and code execution and automations
after that (20-99).

This also skips automations (ocm and sre logins) if launching without
the ocm-container binary, to allow the SRE to decide what to do.

Additionally, this fixes sourcing the AWS cli completions, which are
currently broken in this image, and modifies the KubePS1 config to
display "none" if not logged into OCM, and to always show the current
`ocm config get url` value for the environment, or nothing if unset,
This works even if the sre logs into a new/different one, and updates if
the SRE runs `ocm config set url <some url>`.

This also updates the way the ocm.json file is mounted to match the rest of
the mount options, and will mount it RW, so if the file exists, users
will stay logged into ocm until their token expires, and not require a
device auth for each container.  Environment will change based on the
--ocm-url provided (with production being the default).

Finally, this removes to util scripts that have been broken and out of
date since backplane was introduced, and who's functions have largely
been replaced by OCM and Backplane commands, and fixes a single makefile
target that had incorrect syntax.

In practice this should have little impact directly on ocm-contianer
users, other than they will now be able to switch environments just by
setting the config value after login.  Login still defaults to
production, and can be overridden by the binary with the OCMC_OCM_URL
env variable that is passed in if the user specifies the `--ocm-url`
flag.

Users using the ocm-contianer binary will stay logged into OCM if their
ocm.json file exists outside the container, so device auth is only
required as their token expires.

This is partially in support of
https://issues.redhat.com/browse/OSD-15847, to allow usage of the
ocm-container image elsewhere, and in preparation for changes to the
binary to make device auth login easier.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>