Add 'name' parameter for OCM cluster search
Some clusters use "display_name" as a longer, more descriptive field.
And example of this is `osd-v4stg-aws`, which has:

```
Display Name:	SRE long lived cluster in production: osd-v4stg-aws
```

This causes osd-v4stg-aws (or others) to not be found when searching by the human
friendly name, as identified by @rendhalver in a [recent pr
review](https://github.com/openshift/ocm-container/pull/300#issuecomment-2287870189).

This PR adds "name" to the OCM cluster search parameters, which is the
parameter used by ocm-cli when it performs the same search.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>