Update engine tests to sort strings for testing
The engine string conversion test TestEnvsToString failed due to unsorted appending of the input slices when evaluating the string output.  This updates EnvsToString to sort the input map by keys, and adds an extra test to ensure the sorting is working.

This would not matter in execution, but ensure the expected order for the test evaluation.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>