Various updates for patches/go version (#324)
* Update dependency packages

* Update dependency packages
* Updates OCM pkg to use the ocm-common package rather than ocm directly

Signed-off-by: Chris Collins <collins.christopher@gmail.com>

* Update engine tests to sort strings for testing

The engine string conversion test TestEnvsToString failed due to unsorted appending of the input slices when evaluating the string output.  This updates EnvsToString to sort the input map by keys, and adds an extra test to ensure the sorting is working.

This would not matter in execution, but ensure the expected order for the test evaluation.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>

* Fix fmt make target

Fixes the fmt make target by removing the unused "tests" directory, and adding the "utils" directory for parsing.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>

* Update golangci-lint and goreleaser versions

Updates golangci-lint and goreleaser versions used in the Makefile

Signed-off-by: Chris Collins <collins.christopher@gmail.com>

* Add fixes for golangci-lint

Add golangci-lint fixes

Signed-off-by: Chris Collins <collins.christopher@gmail.com>

* Fix build-snapshot make target name

Fixes a typo in the build-snapshot call in the go_build target

Signed-off-by: Chris Collins <collins.christopher@gmail.com>

* Change name of the go-build target to be consistent

Changes the name of the go-build Make target from go_build, to be consistent with other Make targets in the file

Signed-off-by: Chris Collins <collins.christopher@gmail.com>

* Renaming build_binary make target for consistency

Renaming build_bianry make target to build-binary, to be consistent with other make targets

Signed-off-by: Chris Collins <collins.christopher@gmail.com>

* Use golangci-lint yaml vs toml; std-error-handling preset

* Switches to YAML format for golangci-lint
* Uses exclusion preset for defered *.Close() functions

REF: https://golangci-lint.run/usage/false-positives/#exclusion-presets (std-error-hankdling)

Signed-off-by: Chris Collins <collins.christopher@gmail.com>

---------

Signed-off-by: Chris Collins <collins.christopher@gmail.com>