Add micro/minimal image builds
This PR adds micro and minimal image builds to the Makefile, and some helper functions to make the Makefile more DRY. It also formats and organizes the Makefile, removes unused targets/cleans-up, and updates the release-build.sh CI script to continue to work with the Makefile and simplify the script.

This does not add the micro or minimal targets to the Manifest build yet, and does not support pushing them to the remote repository yet.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>